backend/requirements-dev.txt
# Test dependencies (install alongside requirements.txt)
pytest>=8.0
pytest-asyncio>=0.24
pytest-httpx>=0.35
aiosqlite>=0.20
ruff>=0.15
